feat(WaitSetting): add WaitSetting trait for managing container wait strategies
